refactor(db tests) - moved tests into separate files per entity
- no code changes were made here, just shuffling things around for
  easier reading/finding
